refactor(cqrs): renamed bus_two to core
