+++ crates/cqrs/src/core/body.rs
