+++ crates/cqrs/src/core/handler.rs
