+++ crates/cqrs/src/core/message.rs
