+++ crates/cqrs/src/core/mod.rs
