+++ crates/cqrs/src/core/route.rs
