+++ crates/cqrs/src/lib.rs
@@ -1,2 +1,2 @@
 // mod bus;
-mod bus_two;
+mod core;
